feat: add React Router DOM for post-signup navigation                   - create migrations for tenants, plans, and subscriptions tables                   - create Tenant model and define relationships with User - refactor: rename 'company_name' column to 'tenant_id' in users table                - chore: update UserController signup logic to associate tenant_id

+++ Subscribly/database/migrations/2025_08_19_123559_add_company_name_column_to_users__table.php
@@ -4,17 +4,18 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
+return new class extends Migration {
     /**
      * Run the migrations.
      */
     public function up(): void
     {
-        
-        Schema::table('users', function (Blueprint $table) {
-            $table->string('company_name')->after('email');    
-        });
+        if (!Schema::hasColumn('users', 'company_name')) {
+
+            Schema::table('users', function (Blueprint $table) {
+                $table->string('company_name')->after('email');
+            });
+        }
     }
 
     /**
@@ -23,7 +24,10 @@ return new class extends Migration
     public function down(): void
     {
         Schema::table('users', function (Blueprint $table) {
+           if (Schema::hasColumn('users', 'company_name')) {
+
             $table->dropColumn('company_name');
+           }
         });
     }
 };
